+++ src/classes_body_ships.py
@@ -63,3 +63,16 @@ class Battleship_body(Small_ship):
 
     hit_box_radius = 100
     base_HP = 10000
+
+
+class Aircraft_carrier_body(Small_ship):
+    path = AIRCRAFT_CARRIER_PATH
+    number_of_frames = AIRCRAFT_CARRIER_FRAMES
+    min_scale_to_be_visible = 0.125
+
+    v_max = 0.2
+    acceleration = 0.1
+    turn_speed = 0.005
+
+    hit_box_radius = 100
+    base_HP = 2000