+++ doc/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 ======
 
+v0.16.0 - 11.03.2024
+----
+* New unit: Aircraft_carrier;
+* New fuel consumption mechanics for aircraft.
+
+v0.16 - Planes stage II
+======
+
+
 v0.15.5 - 14.01.2024
 ----
 * Minor changes;